refactor(core): Rename 'ons.js' to 'index.js'.
