+++ core/src/elements/ons-carousel-item.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons/ons';
+import ons from '../ons';
 import util from '../ons/util';
 import BaseElement from './base/base-element';
 import ModifierUtil from '../ons/internal/modifier-util';
