+++ core/src/elements/ons-carousel.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons/ons';
+import ons from '../ons';
 import util from '../ons/util';
 import BaseElement from './base/base-element';
 import contentReady from '../ons/content-ready';
