+++ core/src/elements/ons-splitter-content.js
@@ -15,7 +15,7 @@ limitations under the License.
 
 */
 
-import ons from '../ons/ons';
+import ons from '../ons';
 import util from '../ons/util';
 import internal from '../ons/internal';
 import ModifierUtil from '../ons/internal/modifier-util';
