+++ core/src/index.esm.js
@@ -1,5 +1,5 @@
 import setup from './setup.js'; // Add polyfills -- Must be first
-const ons = require('./ons/ons').default; // 'require' avoids external dependency hoisting
+const ons = require('./ons').default; // 'require' avoids external dependency hoisting
 
 setup(ons); // Setup initial listeners
